Move `verify.SameSubnet` to `subnets.Same`
