+++ subnets/same.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package verify
+package subnets
 
 import (
 	"context"
@@ -17,10 +17,10 @@ var (
 	ErrMismatchedSubnetIDs = errors.New("mismatched subnetIDs")
 )
 
-// SameSubnet verifies that the provided [ctx] was provided to a chain in the
-// same subnet as [peerChainID], but not the same chain. If this verification
-// fails, a non-nil error will be returned.
-func SameSubnet(ctx context.Context, chainCtx *snow.Context, peerChainID ids.ID) error {
+// Same verifies that the provided [ctx] was provided to a chain in the same
+// subnet as [peerChainID], but not the same chain. If this verification fails,
+// a non-nil error will be returned.
+func Same(ctx context.Context, chainCtx *snow.Context, peerChainID ids.ID) error {
 	if peerChainID == chainCtx.ChainID {
 		return ErrSameChainID
 	}
