+++ vms/avm/txs/executor/semantic_verifier.go
@@ -9,6 +9,7 @@ import (
 	"reflect"
 
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/subnets"
 	"github.com/ava-labs/avalanchego/vms/avm/states"
 	"github.com/ava-labs/avalanchego/vms/avm/txs"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
@@ -89,7 +90,7 @@ func (v *SemanticVerifier) ImportTx(tx *txs.ImportTx) error {
 		return nil
 	}
 
-	if err := verify.SameSubnet(context.TODO(), v.Ctx, tx.SourceChain); err != nil {
+	if err := subnets.Same(context.TODO(), v.Ctx, tx.SourceChain); err != nil {
 		return err
 	}
 
@@ -127,7 +128,7 @@ func (v *SemanticVerifier) ExportTx(tx *txs.ExportTx) error {
 	}
 
 	if v.Bootstrapped {
-		if err := verify.SameSubnet(context.TODO(), v.Ctx, tx.DestinationChain); err != nil {
+		if err := subnets.Same(context.TODO(), v.Ctx, tx.DestinationChain); err != nil {
 			return err
 		}
 	}
