+++ vms/platformvm/txs/executor/standard_tx_executor.go
@@ -13,10 +13,10 @@ import (
 
 	"github.com/ava-labs/avalanchego/chains/atomic"
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/subnets"
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/vms/components/avax"
-	"github.com/ava-labs/avalanchego/vms/components/verify"
 	"github.com/ava-labs/avalanchego/vms/platformvm/state"
 	"github.com/ava-labs/avalanchego/vms/platformvm/txs"
 )
@@ -142,7 +142,7 @@ func (e *StandardTxExecutor) ImportTx(tx *txs.ImportTx) error {
 	// Skip verification of the shared memory inputs if the other primary
 	// network chains are not guaranteed to be up-to-date.
 	if e.Bootstrapped.Get() && !e.Config.PartialSyncPrimaryNetwork {
-		if err := verify.SameSubnet(context.TODO(), e.Ctx, tx.SourceChain); err != nil {
+		if err := subnets.Same(context.TODO(), e.Ctx, tx.SourceChain); err != nil {
 			return err
 		}
 
@@ -213,7 +213,7 @@ func (e *StandardTxExecutor) ExportTx(tx *txs.ExportTx) error {
 	copy(outs[len(tx.Outs):], tx.ExportedOutputs)
 
 	if e.Bootstrapped.Get() {
-		if err := verify.SameSubnet(context.TODO(), e.Ctx, tx.DestinationChain); err != nil {
+		if err := subnets.Same(context.TODO(), e.Ctx, tx.DestinationChain); err != nil {
 			return err
 		}
 	}
